*Add reconstruction of sensitivity maps (with_smaps parameter). Avoid batch job submission if file name is empty.
If with_smaps=True in config file, the number of sessions specified in sub_ses is considered double. Every file name MUST be accompanied by an smap_file name in this case. If there is no smap or you don't want to specify one for a specific session, you still need to add an empty string ''.